requirements.txt
# 核心依赖（可选，没有也能用）
Pillow>=9.0.0          # 照片 EXIF 读取

# 以下为可选增强
# chardet>=5.0.0       # 自动检测文件编码
# python-dateutil>=2.8 # 更强的日期解析

openai>=2.32.0
pymilvus>=2.6.12
python-dotenv>=1.2.2
tqdm>=4.67.3
tencentcloud-sdk-python>=3.1.81